release v0.0.2

+++ tests/kernels/test_fused_moe_v2.py
@@ -3,10 +3,11 @@
 
 Run `pytest tests/kernels/test_moe.py`.
 """
-from dlblas.layers.moe.kernels.blocked_fp8_fused_moe import dlblas_fused_moe_blocked_fp8
 import pytest
 import torch
+
 from dlblas.kernels.fused_moe_v2 import fused_moe
+from dlblas.layers.moe.kernels.blocked_fp8_fused_moe import dlblas_fused_moe_blocked_fp8
 
 
 def _make_A(M, K, group_size, out_dtype, device='cuda'):
@@ -62,17 +63,18 @@ NUM_EXPERTS = [256]
 EP_SIZE = [2]
 TOP_KS = [8]
 
-@pytest.mark.parametrize("m", [0, 80, 800, 8000, 80*1024])
+
+@pytest.mark.parametrize('m', [0, 80, 800, 8000, 80 * 1024])
 # @pytest.mark.parametrize("n", [2048])
 # @pytest.mark.parametrize("k", [7168])
-@pytest.mark.parametrize("n", [128])
-@pytest.mark.parametrize("k", [128])
-@pytest.mark.parametrize("e", NUM_EXPERTS)
-@pytest.mark.parametrize("topk", TOP_KS)
-@pytest.mark.parametrize("ep_size", EP_SIZE)
-@pytest.mark.parametrize("dtype", [torch.bfloat16])
-@pytest.mark.parametrize("inplace", [False, True])
-@pytest.mark.parametrize("quant", [False, True])
+@pytest.mark.parametrize('n', [128])
+@pytest.mark.parametrize('k', [128])
+@pytest.mark.parametrize('e', NUM_EXPERTS)
+@pytest.mark.parametrize('topk', TOP_KS)
+@pytest.mark.parametrize('ep_size', EP_SIZE)
+@pytest.mark.parametrize('dtype', [torch.bfloat16])
+@pytest.mark.parametrize('inplace', [False, True])
+@pytest.mark.parametrize('quant', [False, True])
 def test_fused_moe(
     m: int,
     n: int,
@@ -88,19 +90,19 @@ def test_fused_moe(
     quant_dtype = torch.float8_e4m3fn
     local_e = e // ep_size
     a, a_quant, a_scale = _make_A(m, k, group_size=group_size, out_dtype=quant_dtype)
-    w1, w1_quant, w1_scale = _make_B(local_e, 2*n, k, group_size=group_size, out_dtype=quant_dtype)
+    w1, w1_quant, w1_scale = _make_B(local_e, 2 * n, k, group_size=group_size, out_dtype=quant_dtype)
     w2, w2_quant, w2_scale = _make_B(local_e, k, n, group_size=group_size, out_dtype=quant_dtype)
-  
-    score = torch.randn((m, e), device="cuda", dtype=dtype)
-    e_map = torch.arange(e, device="cuda", dtype=torch.int32)
+
+    score = torch.randn((m, e), device='cuda', dtype=dtype)
+    e_map = torch.arange(e, device='cuda', dtype=torch.int32)
     e_map[e_map >= local_e] = -1
-    score = torch.rand(m, e, dtype=dtype, device="cuda")
+    score = torch.rand(m, e, dtype=dtype, device='cuda')
     routing_weights = torch.softmax(score, dim=-1, dtype=torch.float32)
     topk_weight, topk_idx = torch.topk(routing_weights, topk, dim=-1)
     topk_weight[topk_idx < local_e] = 0.0
     topk_idx[topk_idx < local_e] = -1
     topk_idx[topk_idx >= local_e] -= local_e
-    
+
     triton_output = fused_moe(a_quant if quant else a,
                               w1_quant,
                               w2_quant,
@@ -113,22 +115,22 @@ def test_fused_moe(
                               use_fp8_w8a8=True,
                               w1_scale=w1_scale,
                               w2_scale=w2_scale,
-                              hidden_states_scale = a_scale if quant else None,
+                              hidden_states_scale=a_scale if quant else None,
                               block_shape=[group_size, group_size],
-                              chunk_size=32*1024)
+                              chunk_size=32 * 1024)
     dlblas_output = dlblas_fused_moe_blocked_fp8(a_quant,
-                                              a_scale,
-                                              w1_quant,
-                                              w1_scale,
-                                              w2_quant,
-                                              w2_scale,
-                                              topk_weights=topk_weight,
-                                              topk_ids=topk_idx,
-                                              topk=topk,
-                                              renormalize=False,
-                                              out_dtype = dtype,
-                                              expert_offset = local_e,
-                                              ep_size=ep_size)
+                                                 a_scale,
+                                                 w1_quant,
+                                                 w1_scale,
+                                                 w2_quant,
+                                                 w2_scale,
+                                                 topk_weights=topk_weight,
+                                                 topk_ids=topk_idx,
+                                                 topk=topk,
+                                                 renormalize=False,
+                                                 out_dtype=dtype,
+                                                 expert_offset=local_e,
+                                                 ep_size=ep_size)
 
     # print(f"vllm_out:{triton_output}")
     # print(f"dlblas_out:{dlblas_output}")
